refactor unit test examples
