+++ book/smoothing/bracket_morph.idraw
